images: tighten platform resolution structure
